+++ cli/cli-probe-server/src/test/java/io/telicent/smart/cache/cli/probes/TestHealthProbeServer.java
@@ -34,9 +34,13 @@ import org.testng.annotations.AfterClass;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.Test;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Random;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicInteger;
 
 public class TestHealthProbeServer {
@@ -66,7 +70,7 @@ public class TestHealthProbeServer {
         this.server = new HealthProbeServer(null, TEST_PORT.incrementAndGet(), null);
 
         // When
-        server.run();
+        startServer();
 
         // Then
         verifyErrorsLogged();
@@ -78,7 +82,7 @@ public class TestHealthProbeServer {
         this.server = new HealthProbeServer("Test", -100, null);
 
         // When
-        server.run();
+        startServer();
 
         // Then
         verifyErrorsLogged();
@@ -101,9 +105,9 @@ public class TestHealthProbeServer {
                                                                                        "unhealthy")));
     }
 
-    private void verifyLiveness(HealthProbeServer entrypoint) {
+    private void verifyLiveness(HealthProbeServer server) {
         // Given
-        WebTarget target = client.target(entrypoint.getLivenessUrl());
+        WebTarget target = client.target(server.getLivenessUrl());
         Invocation.Builder builder = target.request(MediaType.APPLICATION_JSON);
 
         // When
@@ -113,9 +117,9 @@ public class TestHealthProbeServer {
         Assert.assertNotNull(versionInfo);
     }
 
-    private HealthStatus verifyReadiness(HealthProbeServer entrypoint, boolean expectedReadiness) {
+    private HealthStatus verifyReadiness(HealthProbeServer server, boolean expectedReadiness) {
         // Given
-        WebTarget target = client.target(entrypoint.getReadinessUrl());
+        WebTarget target = client.target(server.getReadinessUrl());
         Invocation.Builder builder = target.request(MediaType.APPLICATION_JSON);
 
         // When
@@ -136,7 +140,7 @@ public class TestHealthProbeServer {
         this.server = new HealthProbeServer("Test", TEST_PORT.incrementAndGet(), null);
 
         // When
-        server.run();
+        startServer();
 
         // Then
         verifyNoErrorsLogged();
@@ -155,7 +159,7 @@ public class TestHealthProbeServer {
                                                               .build());
 
         // When
-        server.run();
+        startServer();
 
         // Then
         verifyNoErrorsLogged();
@@ -163,6 +167,90 @@ public class TestHealthProbeServer {
         // And
         verifyLiveness(server);
         verifyReadiness(server, true);
+    }
+
+    private void startServer() {
+        server.run();
+
+        // As the Health Probe Server runs on a background thread it won't be available immediately so if we proceed to
+        // try and make HTTP requests to it straight away our tests can fail, thus for these tests want a brief wait
+        // after starting the server
+        try {
+            Thread.sleep(250L);
+        } catch (InterruptedException e) {
+            // Ignored
+        }
+    }
+
+    @Test
+    public void givenAlwaysHealthyConfiguration_whenRunningProbeServer_thenSuccess_andServerRespondsToManyRequests() {
+        // Given
+        this.server = new HealthProbeServer("Test", TEST_PORT.incrementAndGet(),
+                                            () -> HealthStatus.builder()
+                                                              .healthy(true)
+                                                              .build());
+
+        // When
+        startServer();
+
+        // Then
+        verifyNoErrorsLogged();
+
+        // And
+        verifyManyHealthProbesSucceed();
+    }
+
+    @Test
+    public void givenSlowReadinessCheck_whenRunningProbeServer_thenSuccess_andServerRespondsToManyRequestsEventually() {
+        // Given
+        this.server = new HealthProbeServer("Test", TEST_PORT.incrementAndGet(),
+                                            () -> {
+                                                try {
+                                                    Thread.sleep(250);
+                                                } catch (InterruptedException e) {
+                                                    // Ignore
+                                                }
+                                                return HealthStatus.builder()
+                                                                   .healthy(true)
+                                                                   .build();
+                                            });
+
+        // When
+        startServer();
+
+        // Then
+        verifyNoErrorsLogged();
+
+        // And
+        verifyManyHealthProbesSucceed();
+    }
+
+    private void verifyManyHealthProbesSucceed() {
+        ExecutorService executorService = Executors.newFixedThreadPool(10);
+        try {
+            List<Future<?>> futures = new ArrayList<>();
+            for (int i = 1; i <= 100; i++) {
+                futures.add(executorService.submit(() -> {
+                    verifyLiveness(this.server);
+                    verifyReadiness(this.server, true);
+                }));
+            }
+            List<Boolean> results = new ArrayList<>();
+            for (Future<?> future : futures) {
+                try {
+                    future.get();
+                    results.add(true);
+                } catch (Throwable e) {
+                    System.err.println(e.getMessage());
+                    results.add(false);
+                }
+            }
+            long failures = results.stream().filter(x -> !x).count();
+            Assert.assertTrue(results.stream().allMatch(x -> x),
+                              "Not all requests successfully completed (" + failures + " failed)");
+        } finally {
+            executorService.shutdownNow();
+        }
     }
 
     @Test
@@ -175,7 +263,7 @@ public class TestHealthProbeServer {
                                                               .build());
 
         // When
-        server.run();
+        startServer();
 
         // Then
         verifyNoErrorsLogged();
@@ -193,7 +281,7 @@ public class TestHealthProbeServer {
                                             () -> null);
 
         // When
-        server.run();
+        startServer();
 
         // Then
         verifyNoErrorsLogged();
@@ -211,7 +299,7 @@ public class TestHealthProbeServer {
                                             () -> {throw new RuntimeException("Failed");});
 
         // When
-        server.run();
+        startServer();
 
         // Then
         verifyNoErrorsLogged();
@@ -236,7 +324,7 @@ public class TestHealthProbeServer {
         });
 
         // When
-        server.run();
+        startServer();
 
         // Then
         verifyNoErrorsLogged();